refactor(console): split main.rs into logging, pid_file and app modules
main.rs had grown to 1800 lines mixing the entry point with logging
setup, PID file handling, config loading, headless mode, subcommand
handlers and R setup. Move each concern into its own module, leaving
main.rs as the entry point plus run(). Pure code movement, no behavior
change.
